added models, repository and service- Reformated, moved to dto package.
