+++ Users-management-ws/src/main/java/com/silverlining/usermanagement/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.silverlining.usermanagement.model;
+package com.silverlining.usermanagement.dto;
 
 public class UserDto {
     String id;
