+++ Users-management-ws/src/main/java/com/silverlining/usermanagement/repository/UserRepository.java
@@ -1,11 +1,9 @@
 package com.silverlining.usermanagement.repository;
 
 import com.silverlining.usermanagement.model.User;
-import com.silverlining.usermanagement.model.UserDto;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
